Flattend out G::territoryMap; custom rand function (std::vector<std::vector<Pixel*>> => std::vector<Pixel*>)

+++ standard/src/map/TerrainKind.h
@@ -44,7 +44,7 @@ namespace Terrain {
 
     Color GetColor(Kind kind);
 
-    float GetLevel(Kind kind);
+    int GetLevel(Kind kind);
 
     float GetInvasionProbability(Kind kind);
